+++ packages/payload/seeder/src/types/tokens/seedTokens.ts
@@ -1,0 +1,9 @@
+import type {
+  file,
+  ref,
+} from '#refs'
+
+export interface SeedTokens {
+  file: typeof file
+  ref: typeof ref
+}