+++ packages/api/api-error/CHANGELOG.md
@@ -1,6 +1,12 @@
 # @wisemen/api-error
 
 
+
+## 1.1.0
+<sub>2026-07-29</sub>
+
+- [#1518](https://github.com/wisemen-digital/wisemen-core/pull/1518)  *(minor)* Thanks [@Kobe-Kwanten](https://github.com/Kobe-Kwanten)! - add class-validator json api error conversion
+
 ## 1.0.1
 <sub>2026-07-28</sub>